moved user authentication into its own file and added try/except in case of failure
