+++ backend/urls.py
@@ -1,20 +1,9 @@
-
 from core.models import *
 from core.middleware import *
 from django_instant_rest import patterns
 from django.urls import path
+from .views import authenticate
 
-from django.http import HttpResponse
-from django.views.decorators.csrf import csrf_exempt
-import json
-
-@csrf_exempt
-def authenticate(request):
-    credentials = json.loads(request.body.decode("utf-8"))
-    c = Contributor.objects.get(username=credentials["username"])
-    result = c.verify_password(credentials["password"])
-    print(result)
-    return HttpResponse(result)
 
 urlpatterns = [
     patterns.resource('issues', Issue, middleware = issue_resource_policies),
